[recipes] Define modal transformers
